organize network classes
organized files into corresponding directories
