+++ app/src/main/java/com/PrivacyGuard/Application/Network/Protocol/IP/IPv6Header.java
@@ -1,4 +1,4 @@
-package com.PrivacyGuard.Application.Network.IP;
+package com.PrivacyGuard.Application.Network.Protocol.IP;
 
 import com.PrivacyGuard.Utilities.ByteOperations;
